requirements.txt
absl-py==1.4.0
anyio==3.7.0
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
DBUtils==3.0.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dnspython==2.2.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.10.0
Flask==2.2.2
Flask-APScheduler==1.12.4
flatbuffers==23.5.9
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.6.0
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.4
iniconfig==1.1.1
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lightgbm==3.3.3
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.1.0
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
nodeenv==1.7.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.3.1
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pip-tools==6.14.0
platformdirs==2.5.4
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.36
protobuf==4.23.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.2
PyMySQL==1.0.2
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.2
sentencepiece==0.1.99
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
torch==2.0.0
torchinfo==1.7.2
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
transformers==4.30.2
typing_extensions==4.5.0
tzdata==2022.5
tzlocal==4.2
uri-template==1.3.0
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.2
wrapt==1.14.1
